Final commit?
I am not sure if this commit will fail or not. Hope it works! It is just
kind of everything finally put into one or two places, really just
everything into ActuallyMain.java If we just change a couple lines of
code half of it works and if we change it back the other half works.
It's fine.